+++ R/mosaic.R
@@ -304,6 +304,11 @@ mosaic <- function(data = NULL,
   ## filter out base levels if group defined
   prodcoords <- prodcoords[stats::complete.cases(prodcoords), ]
 
+  if (use_facet) {
+    prodcoords$facet <- factor(prodcoords$facet,
+      levels = levels(data[, symlist[["facet"]]])
+    )
+  }
   ## if group sizes, then calculate label coords
   if (group_counts) {
     prodcoords[, "size_x"] <- (prodcoords[, "l"] + prodcoords[, "r"]) / 2